- Get our Win32 version of ccache to compile warning-free under MSVC (tested with 2008 at /W4) and GCC (tested with RosBE-Windows 1.2 at -Wall -pedantic) - Fix the MSVC project file - Add a Makefile for compiling with mingw32-gcc
git-svn-id: file:///home/aleksey/repos/reactos_mirr/trunk@34306 9874b476-3cf2-4131-8f6d-dcf341c5f767

+++ tools/RosBE/RosBE-Windows/Tools/ccache-2.4/win32-dirent.c
@@ -73,7 +73,7 @@ readdir (DIR * dir)
 
   entry.d_off = dir->filepos;
   strncpy (entry.d_name, find->cFileName, sizeof (entry.d_name));
-  entry.d_reclen = strlen (find->cFileName);
+  entry.d_reclen = (unsigned short)strlen (find->cFileName);
   dir->filepos++;
   return &entry;
 }
